3._DateServers_Part1_&_Part2/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Gabriel <gabriel_nicutari@yahoo.ro>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.81.0"
uvicorn = "^0.18.3"
requests = "^2.28.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
